+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Blacklist slot and bank slot.
 - You can now open the bank and auction windows from the Manager inventory tab.
 - LootUI toggle hotkey can now be customized in the settings panel.
+- A hotkey to toggle autoloot on and off has been added. This can be customized in the settings panel.
 
 ### Fixed
 - Flag check to prevent looting NPCs that should not normally be able to be looted.